+++ src/Campaigns/resources/admin/components/CampaignDetailsPage/CampaignDetailsPage.module.scss
